Move all execution code to Schema class
