+++ pyproject.toml
@@ -96,6 +96,7 @@ types-deprecated = "^1.2.15.20241117"
 types-six = "^1.17.0.20241205"
 types-pyyaml = "^6.0.12.20240917"
 mypy = "^1.13.0"
+codeflash = "^0.9.2"
 
 [tool.poetry.group.integrations]
 optional = true
