+++ strawberry/extensions/validation_cache.py
@@ -3,7 +3,6 @@ from functools import lru_cache
 from typing import Optional
 
 from strawberry.extensions.base_extension import SchemaExtension
-from strawberry.schema.execute import validate_document
 
 
 class ValidationCache(SchemaExtension):
@@ -32,6 +31,8 @@ class ValidationCache(SchemaExtension):
 
         More info: https://docs.python.org/3/library/functools.html#functools.lru_cache
         """
+        from strawberry.schema.schema import validate_document
+
         self.cached_validate_document = lru_cache(maxsize=maxsize)(validate_document)
 
     def on_validate(self) -> Iterator[None]:
