+++ tests/schema/test_execution.py
@@ -11,7 +11,7 @@ from strawberry.extensions import AddValidationRules, DisableValidation
 
 
 @pytest.mark.parametrize("validate_queries", [True, False])
-@patch("strawberry.schema.execute.validate", wraps=validate)
+@patch("strawberry.schema.schema.validate", wraps=validate)
 def test_enabling_query_validation_sync(mock_validate, validate_queries):
     @strawberry.type
     class Query:
@@ -64,7 +64,7 @@ async def test_enabling_query_validation(validate_queries):
         }
     """
 
-    with patch("strawberry.schema.execute.validate", wraps=validate) as mock_validate:
+    with patch("strawberry.schema.schema.validate", wraps=validate) as mock_validate:
         result = await schema.execute(
             query,
             root_value=Query(),
